Implement remote snapshot.
This relies on infrastructure added by PoWA 4.

The collector now get the list of datasources from the repository for the given
instance, export each datasource's data from the remote instance to a
repository table, and call powa_take_snapshot for the instance.

Some basic logging and reporting is also added.