Properly handle snapshot time at collector/thread restart
Each thread now retrieve the previous snapshot time stored on the database.

If it's older than configured frequency, choose a random point between now and
now minus interval to spread the snapshots and avoid activity spikes if the
collector itself was stopped for a long time, or if a lot of new servers were
added.