Better handling of lack of powa extension on remote servers.
As reported by Thoma Reiss, if the extension isn't available on the remote
server, users get an error message that's not helpful at all.  To fix, first
do the sanity check for the extension in the catalog to possibly avoid to hit
an error when trying to load powa, and also add a specific error message if the
extension cannot be loaded.