Don't perform a commit after executing the cleanup query.
The snapshot assumption, whether local or remote, relies on everything being
done in a single transaction, and more precisely having now() returning a
unique timestamp.  Adding a commit in the middle on the remote snapshot process
breaks this assumption, leading to half of the UI being broken as Adrien Nayra
reported.