Close all thread's connections in case of uncatched error during snapshot.
Errors during per-module snapshot should be properly handled, but if any error happens outside of it, we can still raise an exception.  Those errors shouldn't be frequent and should most of the time be the sign of a bigger problem.  So simply close both repository and remote server connections in that case, to make sure that the next snapshot won't be working with leftover problems of the previous snapshot attempt.