Automatically detect hypopg on remote server.
And save the default version in the repository powa_extensions table.

It doesn't mean that the extension will installed and usable when needed on
each remote database, but at least users can confirm that the extension is
available on the remote server on the powa-web server configuration page.

Thanks to MikhailSaynukov for the report.