Use proper bind param for notification payload response.
Currently, the generated payload is entirely controlled by powa-collector, so
there's no injection risk or any other problem.

A later commit will introduce messages for which the payload may depend on user
input, so we will need to make sure the values are properly escaped.  The
channel name is still totally controlled so there's no problem inserting it
as-is.  Identifiers can't use parameters, so we don't have a choice anyway.

While at it, generate a better response when the command isn't understood,
which also depends on user input, to help debugging any issue in the calling
code.