Add a function to retrieve the worker thread for a given remote server.
An exception is raised if there's no worker for this server, meaning that it's
an unknown server (or the remote server is stopped, or the worker couldn't
connect, but we don't try to distinguish those cases).

This function will be needed by a later commit for processing notifications, so
also change the loop around __process_one_notification() to handle exceptions.