Add a FORCE_SNAPSHOT notification message.
The full syntax is

FORCE_SNAPSHOT $srvid

where $srvid is an integer.

This allows callers to request an immediate snapshot for the given remote
server.  To avoid doing too many snapshots, a new snapshot will be forced only
if there isn't an ongoing snapshot already and if the last snapshot finished
more than 10s ago.

Note that there is no attempt to handle any form of authorization.  This will
be done in a later commit.