WIP: Add a REFRESH_DB_CAT notification message.
The full syntax is

REFRESH_DB_CAT $srvid, $nb_databases [$dbname ...]

where $srvid is an integer, $nb_databases is an integer, and $dbname an
optional list of space-separated database names.

This allows callers to ask for the catalogs to be refreshed for the given
remote server, possibly on a subset of databases, at the next snapshot.

A mismatch between $nb_databases and the number of provided database names will
return an error.  We require the number of database to make sure that we didn't
received a truncated message.