fix: Properly detect and report missing extensions (#32)
Doing an extension remote snapshot relies on the extension being installed
both on the remote server but also on the repository server.  If the extension
is missing on either side the collector will currently fail with an obscure
error about a missing key when trying to resolve the schema the extension is
installed in.  We now properly check whether we found a schema for the wanted
extension, and if not report a proper error.

Thanks to github users jw1u1 and rafaelorafaelo who both reported the problem.